fix library warnings
Clears the four warnings emitted by jill/ now that warnings are on in
every build configuration.

- mirrored_memory: cast SHMLBA before comparing it against a size_t.
  Annotation only; SHMLBA is always positive.

- triggered_data_writer: the flush sanity check kept a variable alive
  solely to be asserted on, leaving a dead read under NDEBUG. peek() is
  const and side-effect free, so the check moves inside the asserts and
  compiles out entirely. Split in two so a failure identifies which
  invariant broke.

- zmq: zmq_msg_init() is documented to always return zero, so the check
  on its result was vacuous.

- zmq: zmq_msg_init_size() can fail with ENOMEM, and the assert guarding
  it disappeared under NDEBUG, so a release build could return a message
  that was never initialized. The message is also now initialized before
  being handed to the shared_ptr, so the msg_close deleter can never run
  against an uninitialized zmq_msg_t on the failure path.

+++ jill/util/mirrored_memory.cc
@@ -32,7 +32,9 @@ mirrored_memory::mirrored_memory(size_t arg_size, size_t guard_pages, bool lock_
 {
         int shm_id;
         size_t page_size = getpagesize();
-        if (SHMLBA > page_size)
+        // SHMLBA is a signed int, but is always positive; cast it so this is
+        // not a signed/unsigned comparison.
+        if (static_cast<size_t>(SHMLBA) > page_size)
                 page_size = SHMLBA;
         size_t guard_size = guard_pages * page_size;
